mcp_servers/walmart/requirements.txt
mcp[cli]>=1.12.0
requests
python-dotenv==1.0.0
uvicorn[standard]
starlette
click